Refactor service and repository structure
